+++ .settings/org.eclipse.wst.common.component
@@ -1,28 +1,34 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-        	
+            	
     
-    <wb-module deploy-name="basic-1.0.0-BUILD-SNAPSHOT">
-                        
+    
+    <wb-module deploy-name="odc-1.0.0-BUILD-SNAPSHOT">
+                                
+        
         
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
-                		
+                        		
+        
         
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-                		
+                        		
+        
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
-                		
+                        		
+        
         
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-                		
+                        		
+        
         
         <property name="java-output-path" value="target/classes"/>
-                        
-        
-        <property name="context-root" value="basic"/>
-                	
+        <property name="context-root" value="odc"/>
+                        	
+    
     
     </wb-module>
-        
+            
+
 
 </project-modules>
