+++ src/main/java/com/giltae/main/MainServiceImp.java
@@ -1,4 +1,4 @@
-package com.odc.main;
+package com.giltae.main;
 
 import org.springframework.stereotype.Service;
