+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -21,7 +21,7 @@
 		<beans:property name="suffix" value=".jsp" />
 	</beans:bean>
 	
-	<context:component-scan base-package="com.odc.member, com.odc.main" />
+	<context:component-scan base-package="com.giltae.member, com.giltae.main" />
