+++ src/main/webapp/WEB-INF/views/include/Header.jsp
@@ -21,33 +21,32 @@
 	
 	<div id="headerRight">
 		<ul class="flex-container">
-			<c:choose>
-				<c:when test="${not empty uId }">
-					<c:choose>
-						<c:when test="${uLevel eq '3'}">
-							<li><a href="/adminPage">${uName }님 안녕하세요.</a></li>
-							<li></li>
-							<li><a href="/logout">로그아웃</a></li>                				
-						</c:when>
-						
-						<c:when test="${uLevel eq '2'}">
-							<li><a href="/myPage">${uName }선생님 안녕하세요.</a></li>
-							<li></li>
-							<li><a href="/logout">로그아웃</a></li>
-						</c:when>
-						
-						<c:otherwise>
-							<li><a href="/myPage">${uName }님 환영합니다.</a></li>
-							<li></li>
-							<li><a href="/logout">로그아웃</a></li>
-							<li></li>
-							<li>
-							<div id="headerCart">
-							<a href="/cart.jsp"><img src="${path }/resources/img/cart.png" alt=""><span><%-- <%=cartCnt %> --%>0</span></a>
-							</div>
-							</li>
-						</c:otherwise>
-					</c:choose>
+		
+<%--  			<c:choose>
+				<!-- uLevel -> 3번 관리자등급 -->
+				<c:when test="${not empty uId && uLevel eq '3'}">
+					<li><a href="/adminPage">${uName}님 안녕하세요.</a></li>
+					<li></li>
+					<li><a href="/logout">로그아웃</a></li>                				
+				</c:when>
+				
+				<!-- uLevel -> 2번 관리자등급 -->
+				<c:when test="${not empty uId  && uLevel eq '2'}">
+					<li><a href="/myPage">${uName}선생님 안녕하세요.</a></li>
+					<li></li>
+					<li><a href="/logout">로그아웃</a></li>
+				</c:when>
+				
+				<c:when test="${not empty uId}">
+					<li><a href="/myPage">${uName}님 환영합니다.</a></li>
+					<li></li>
+					<li><a href="/logout">로그아웃</a></li>
+					<li></li>
+					<li>
+						<div id="headerCart">
+							<a href="/cart.jsp"><img src="${path}/resources/img/cart.png" alt=""><span><${cartCnt}</span></a>
+						</div>
+					</li>
 				</c:when>
 				
 				<c:otherwise>
@@ -55,7 +54,45 @@
 					<li></li>
 					<li><a href="/member">Sign Up</a></li>
 				</c:otherwise>
+				
+			</c:choose> --%>
+			
+   			<c:choose>
+				<c:when test="${not empty uId }">
+					<c:choose>
+						<c:when test="${uLevel eq '3'}">
+							<li><a href="/adminPage">${uName}님 안녕하세요.</a></li>
+							<li></li>
+							<li><a href="/logout">로그아웃</a></li>                				
+						</c:when>
+						
+						<c:when test="${uLevel eq '2'}">
+							<li><a href="/myPage">${uName} 선생님 안녕하세요.</a></li>
+							<li></li>
+							<li><a href="/logout">로그아웃</a></li>
+						</c:when>
+						
+						<c:otherwise>
+							<li><a href="/myPage">${uName}님 환영합니다.</a></li>
+							<li></li>
+							<li><a href="/logout">로그아웃</a></li>
+							<li></li>
+							<li>
+							<div id="headerCart">
+							<a href="/cart.jsp"><img src="${path }/resources/img/cart.png" alt=""><span>${cartCnt }</span></a>
+							</div>
+							</li>
+						</c:otherwise>
+					</c:choose>
+				</c:when> 
+				
+				<c:otherwise>
+					<li><a href="/login">Sign In</a></li>
+					<li></li>
+					<li><a href="/member">Sign Up</a></li>
+				</c:otherwise>
 			</c:choose>
+				
 		</ul>
 	</div>
 	<!-- div#headerRight -->
